third_party/blink/web_tests/external/wpt/html/semantics/scripting-1/the-script-element/module/dynamic-import/dynamic-imports-credentials-setTimeout.sub-expected.txt

This is a testharness.js-based test.
[FAIL] Cross-origin dynamic import from setTimeout(string) from classic script (crossOrigin=use-credentials)
  assert_equals: cookie expected "found" but got "not found"
[FAIL] Cross-origin dynamic import from setTimeout(string) from module script (crossOrigin=use-credentials)
  assert_equals: cookie expected "found" but got "not found"
Harness: the test ran to completion.
